Update utils.R
The analysis options may contain formulas that are currently not "sourcified" (an empty string is returned). The additional code carries out this transformation - the formula is first converted into a string and then enclosed with "as.formula()". The second condition in else if (! is.symbol(object)) is required because other parameters (data, vars, group, etc.) are originally defined as symbols too.